add test server (#43)
* 🔧: Update config.json formatting for better readability

Reformatted JSON arrays for languages and apiFiles to improve readability and maintainability. Also reordered the Osmosis-lenses group in the navbar configuration.

* ♻️: Sort directories in API menu generation

Added sorting functionality to the API menu generation script to ensure directories are ordered based on their index in the apiConfig. This improves the consistency of the generated sidebars.

* ✨: Add new script for adding docs server

Created a new script file for adding a documentation server, setting the stage for future enhancements and integrations.

* ✨ Add 'servers' field to OpenAPI paths

* ✨ Implement addDocsServer function to append documentation server URLs

* 🔧: Update configuration formatting and add new subpage

Reformatted JSON arrays for languages and apiFiles for consistency. Added a new subpage 'api/osmosis-overview' under the Osmosis group.

* 🔧 Update JSON configuration formatting

Reformatted JSON arrays for better readability and consistency.

* ✨ Add subpages support to API configuration

Introduced optional `subpages` property in `ApiSection` type and updated related logic to handle subpages in API menu generation.

* 🔧 Update configuration structure for Osmosis API

* 🐛 Fix sidebar generation by spreading subpages array